tests/ images service unit test refactor/ controller business logic moved to images service.

+++ src/images/images.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Injectable,
   InternalServerErrorException,
   NotFoundException,
@@ -6,12 +7,14 @@ import {
 import { v2 as cloudinary } from 'cloudinary';
 import { ConfigService } from '@nestjs/config';
 import { PrismaService } from '../prisma/prisma.service';
+import { ProductsService } from '../products/products.service';
 
 @Injectable()
 export class ImagesService {
   constructor(
-    private configService: ConfigService,
-    private prisma: PrismaService,
+    private readonly configService: ConfigService,
+    private readonly prisma: PrismaService,
+    private readonly productsService: ProductsService,
   ) {
     cloudinary.config({
       cloud_name: configService.get<string>('CLOUDINARY_CLOUD_NAME'),
@@ -25,10 +28,22 @@ export class ImagesService {
     mime_type: string,
     product_variation_id: string,
   ) {
+    if (!mime_type.startsWith('image/')) {
+      throw new BadRequestException('Image must have a valid image mime type');
+    }
+
+    const record = await this.productsService.GetProductVariationById(
+      product_variation_id,
+      true,
+    );
+    if (record === 0) {
+      throw new NotFoundException('Product Variation not found');
+    }
+
     const image = await cloudinary.uploader.upload(
       `data:${mime_type};base64,${imageBase64}`,
     );
-    if (!image.asset_id) {
+    if (!image.public_id) {
       throw new InternalServerErrorException(
         'There was an issue with the CDN, please try again later.',
       );
@@ -51,20 +66,22 @@ export class ImagesService {
   }
 
   async deleteImageById(id: string) {
-    const record = await this.prisma.images.findUnique({ where: { id: id } });
+    const record = await this.prisma.images.findUnique({
+      where: { id: id },
+    });
     if (!record) {
       throw new NotFoundException('Image not found');
     }
 
     const cdn_response = await cloudinary.uploader.destroy(record.cdn_id);
 
-    if (cdn_response.result && cdn_response.result === 'ok') {
+    if (cdn_response && cdn_response.result && cdn_response.result === 'ok') {
       return this.prisma.images.delete({
         where: { id: record.id },
       });
     } else {
       throw new InternalServerErrorException(
-        `An error occurred when deleting the image: ${cdn_response}`,
+        `An error occurred when deleting the image.`,
       );
     }
   }
